backend/requirements.txt
psycopg[binary,pool]~=3.2.3
pika~=1.3.2
fastapi~=0.115.6
pydantic~=2.10.3
python-dotenv~=1.0.1
structlog
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-psycopg
opentelemetry-instrumentation-pika
opentelemetry-exporter-otlp-proto-http
prometheus-client
pydantic-settings
uvicorn
python-multipart
sympy~=1.13.1
antlr4-python3-runtime==4.11.0
ollama